refactor(fixture): split query-related stuff out into separate module
